### 2021-03-05 e5b888d
update links to third-party libs' license files
