### 2021-03-11 #3088 @kbarnhart
Thanks for that information @mandli. I'm hoping to get reviewers who both know Clawpack well (e.g., you or @dgeorge) as well as reviewers who are unfamiliar with it. Presuming you are not disqualified by the JOSS Conflict of Interest policy, I'd be interested in your review given your experience with the most recent versions of Clawpack. I think @dgeorge may also be willing to review, such that reviewers might include two people familiar with Clawpack. I'd then find one additional reviewer.

I can hold off on making a final decision about reviewers until I hear from others.
