### 2021-03-15 #3088 @dlgeorge
Hi Katy,

yes, I'm willing to review. Thanks,

dave

On Tue, Mar 9, 2021 at 3:29 PM Katy Barnhart ***@***.***>
wrote:

> 👋 @mandli <https://github.com/mandli> @dlgeorge
> <https://github.com/dlgeorge> @jmansour <https://github.com/jmansour>
> @bzindovic <https://github.com/bzindovic> would you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission is "geoclaw-landspill: an oil land-spill and overland flow
> simulator for pipeline rupture events" and I'm contacting you based on your
> experience with Clawpack and/or fluid flows, as well as Python, and Fortran.
>
> This is a pre-review issue which is used to find reviewers. Once 2-3
> reviewers have been found I'll start the review on a dedicated GitHub
> issue. At present we are asking reviewers to complete reviews in 6 weeks.
> If you are not able to review and have someone to recommend, please mention
> them here (when mentioning, please place a space after the @ of a github
> handle, for example, you would refer to me as "@ kbarnhart").
>
> If you are interested in reviewing, I would recommend looking over the
> journal's conflict of interest policy
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy>
> before the review process starts.
>
> If you have any questions about the JOSS review process, please do not
> hesitate to reach out to me on this issue or at ***@***.***
>
