### 2021-03-15 #3114 @piyueh
Hi @kbarnhart @dlgeorge @jmansour @bzindovic, thanks in advance for reviewing our submission. I want to provide extra information that may help to examine the length of history of our work (one of the review items under _substantial scholarly effort_):

We developed the Fortran code initially in another repository at [https://github.com/barbagroup/geoclaw](https://github.com/barbagroup/geoclaw/tree/land-spill/). However, we recently combined the Fortran code and Python frontend into one repository, i.e., the [geoclaw-landspill](https://github.com/barbagroup/geoclaw-landspill) repository in this submission. So the history of the Fortran code in _geoclaw-landspill_ may look short. If interested, reviewers can find all actual development history (issues, PRs, and code reviews) of the Fortran code in its [original repository](https://github.com/barbagroup/geoclaw/tree/land-spill/).
