### 2021-04-12 #3114 @bzindovic
Dear Katy,

I apologize for not finishing my duties as promised: my 1.5-year-old-son
got a rotavirus infection a few weeks ago and, due to my wife's job
obligations I was parenting the whole time. He was feeling well for a few
days but, as he lost almost 10% of body weight and didn't recuperate fully,
he got a fever and I again had to take care of him. I had very little time
to devote to review so I couldn't finish it as promised. I'll do my best to
complete it as soon as possible.

Regards,
Budo Zindovic

On Mon, Apr 12, 2021 at 5:13 AM Katy Barnhart ***@***.***>
wrote:

> @jmansour <https://github.com/jmansour> @bzindovic
> <https://github.com/bzindovic> @dlgeorge <https://github.com/dlgeorge> a
> friendly reminder that we are about four weeks into this review. JOSS is
> asking reviewers to try and complete their reviews within six weeks. Of
> course is this is not possible, please just keep me posted here. And if you
> have any questions as you work on your reviews, don't hesitate to ask.
>
