### 2021-04-22 #3114 @bzindovic
Dear all,

I've finished review of the paper and I'd like to congratulate the authors on their effort. Some things could be improved as they might be more visible to first-timers:
- [ ] consider adding a link to Contributing.md inside Readme.md. This might be beneficial as the Github page displays contents of Readme.md by default.
- [ ] consider adding a dedicated conda package for your solver. This might simplify the installation of your package.
- [ ] consider adding an explicit statement that Windows is not supported (or that Windows used can install it through WSL).

 A good example that covers all of these is this [JOSS paper](https://joss.theoj.org/papers/10.21105/joss.02696). These recommendations are _not essential_ and I think paper **can be accepted as is**.
