### 2021-04-22 barbagroup/geoclaw-landspill-cases#17 @piyueh
Work on JOSS review by @bzindovic
Refer to this comment by @bzindovic in https://github.com/openjournals/joss-reviews/issues/3114#issuecomment-824697864_
