### 2021-04-22 e00ab80
fix #18: only test plot pixel values against matplotlib v3.3.3
