### 2021-04-27 barbagroup/geoclaw-landspill-cases#20 @jmansour
Minor runtime issues
Hi guys

Just a few minor suggestions as a follow up on the JOSS review process.

1. On my system (Ubuntu 20.04 via Windows WSL), the runtime thread count defaults to the _logical_ processors count, instead of the physical core count. That's effectively oversubscribing the system, as hyper-threading is generally not useful for these problem classes.  If I instead explicitly set `OMP_NUM_THREADS=4` (the number of physical cores), the job runs around 25% faster when compared to the same job without setting `OMP_NUM_THREADS` (which defaults to 8 threads).  I actually prefer as a default that a single thread be used, with the user then required to explicitly configure their environment to use more threads. But otherwise using the physical core count is also a reasonable default.
2. Is there a runtime parameter that adjusts the standard outputting verbosity level? The default the outputting seems somewhat excessive.
