### 2021-04-27 barbagroup/geoclaw-landspill-cases#20 @piyueh
Hi @jmansour, thanks for the review! It's great to hear this project also runs under WSL! We don't have much experience with WSL, so it's great to know this.

1. Yes, from our experience, most of the time it's better to use only physical cores. I think we can set the default OpenMP thread number to the number of physical cores in the Python launcher `geoclaw-landspill`. And users can control the number of threads through a CMD parameter if they want. I'll leave this issue open until we add this feature.

2. Currently, the verbosity of log output is set in the `setrun.py` script in each simulation case. The corresponding parameter is `rundata.clawdata.verbosity`. By default it's 5, meaning all information of AMR mesh of level 5 and lower will be printed out. Changing it to 0 disables the logging, and changing it to 1 only prints the info of the coarsest mesh. I think it's nice to have a CMD parameter in `geoclaw-landspill` to control the verbosity as well. Then there's no need to edit `setrun.py` just to change the log verbosity. Will leave this issue open until done.
