### 2021-05-01 #3114 @dlgeorge
dlgeorge review: geoclaw-landspill
=================

***Review by david george***

Thanks to the authors, I have enjoyed looking at the cases --- very interesting applications. (You could probably apply this to other applications too with some alternative friction models? (lava from a vent?))

## Minor overarching comments

* optional suggestion: "*full* shallow water equations?" Maybe omit "full," or if you want to be more descriptive use "nonlinear shallow water equations" or "2D nonlinear shallow water equations"..."over topography?" (last one is probably obvious and too much though). (Unless *full* is commonly used in this application field?)

* "unnecessary AMR code" is really just refinement criteria...in flag2refine, correct?

* maybe deps_install_tests.md and README.md have...somewhat redundant...but not completely the same info? I'm not sure what the standard protocol is for such...but possibly could be improved...maybe it could just all be in README?

* agree with @bzindovic: "consider adding a dedicated conda package for your solver," though I personally don't know what details that effort would entail.

* I guess that's it, these are just minor optional suggestions.

## Notes about installing/running

* summary: the main issues I had involved getting my own system(s) up to date...some issues with gfortran 8 installation with conda...but this is not a critique of authors' submission obviously...but out-of-date environment required some work. Once these issues were resolved...worked to run...check.

* I gave up on pytest, but again, that's my own envrionment problems I think.

* attempted to run on mac too (authors' clearly state that only linux supported, not a critique)...failed but will keep tweaking & trying and let you know if I figure out anything.
