### 2021-05-05 barbagroup/geoclaw-landspill-cases#21 @piyueh
JOSS review by @dlgeorge
The following items are copied from https://github.com/openjournals/joss-reviews/issues/3114#issuecomment-830489203

- [x] optional suggestion: "full shallow water equations?" Maybe omit "full," or if you want to be more descriptive use "nonlinear shallow water equations" or "2D nonlinear shallow water equations"..."over topography?" (last one is probably obvious and too much though). (Unless full is commonly used in this application field?)

- [x] "unnecessary AMR code" is really just refinement criteria...in flag2refine, correct?

- [x] maybe deps_install_tests.md and README.md have...somewhat redundant...but not completely the same info? I'm not sure what the standard protocol is for such...but possibly could be improved...maybe it could just all be in README?

- [x] agree with @bzindovic: "consider adding a dedicated conda package for your solver," though I personally don't know what details that effort would entail.
