### 2021-05-05 #3114 @piyueh
> # dlgeorge review: geoclaw-landspill
>
> ## Notes about installing/running
> * summary: the main issues I had involved getting my own system(s) up to date...some issues with gfortran 8 installation with conda...but this is not a critique of authors' submission obviously...but out-of-date environment required some work. Once these issues were resolved...worked to run...check.
> * I gave up on pytest, but again, that's my own envrionment problems I think.
> * attempted to run on mac too (authors' clearly state that only linux supported, not a critique)...failed but will keep tweaking & trying and let you know if I figure out anything.

Thank you, @dlgeorge ! We will track and address the issue at https://github.com/barbagroup/geoclaw-landspill/issues/21.

I would like to address the comment regarding the installation. I think that's one reason for having Docker and Singularity containers. As long as users successfully install either Docker or Singularity, they can use this package's Docker images and Singularity images without installing anything else. For Linux, Docker and Singularity are available in major distributions (Ubuntu, Arch, etc.) unless the distro version is too old. For Windows, installing Docker is straightforward with the official installer ([here](https://docs.docker.com/docker-for-windows/install/)). Not sure about Mac. If you would like to try Docker/Singularity images and encounter issues, please let us know!
