### 2021-05-21 a8621ba
address/fix #17 and openjournals/joss-reviews#3114#issuecomment-824697864
