+++ doc/deps_install_tests.md
@@ -1,35 +1,32 @@
 # Dependencies, installation, and tests
 
-The only operating system supported is Linux. We are not maintaining
+The only operating system officially supported is Linux. We are not maintaining
 compatibility with other systems, though they may still work.
 
 ---------------
 ## Dependencies
 
-`pip` can install all except one dependency. File `requirements.txt` describes
-runtime dependencies, and file `requirements-build.txt` describes additional
-dependencies if building the package locally. When installing
-*geoclaw-landspill* through `pip`, `pip` takes care of dependencies
-automatically.
+Build time and runtime dependencies are described in `requirements-build.txt`
+and `requirements.txt`, respectively. `gfortran >= 8.0` is the only build time
+dependency not included in the file, and, correspondingly, `libgfortran5 >= 8.0`
+is the only runtime dependency not included in the file.
 
-The only dependency that may need manual installation is the Fortran compiler.
-We have only tested `gfortran`. After installation, users can remove the
-compiler itself (i.e., `gfortran`). However, `libgfortran` may be needed during
-runtime if using dynamic linking. For `gfortran` (and also `libgfortran`), the
-minimum version is 8.
+Anaconda users do not need to worry about dependencies at all.
 
-Installation of `gfortran` depends on the operating systems and virtual
-environments. For example, in Arch Linux:
+On the other hand, `pip` users have to install `gfortran` or/and `libgfortran5`
+in advance using the package managers of their Linux distributions. For example,
+in Arch Linux, use:
 ```
 # pacman -S gcc-fortran
 ```
-Ubuntu 20.04:
+And in Ubuntu 20.04:
 ```
 # apt install gfortran
 ```
 
-Alternatively, one can use `conda` to get `gfortran` if using Anaconda virtual
-environment:
+Alternatively, though not recommended, one can use `conda` to get `gfortran` and
+then continue using `pip` for other dependencies. The command to get `gfortran`
+from Anaconda is
 ```
 $ conda install -c conda-forge "gfortran_linux-64>=8.0"
 ```
@@ -37,23 +34,44 @@ $ conda install -c conda-forge "gfortran_linux-64>=8.0"
 However, this should not concern users because CMake should be able to find it
 automatically.
 
+After installing `gfortran` manually, `pip` users can continue on the
+installation of *geoclaw-landspill* (the next section).
+
 ---------------
 ## Installation
 
-Note, for the `pip` command and `python setup.py install`, users can always add
-the `--user` flag to install to users' local paths and avoid root privilege.
+### Option 1: use `conda` and install binary files from Anaconda
 
-### 1. Install from PyPI
+As described in README, the following command creates an environment called
+`landspill`, and it has *geoclaw-landspill* installed:
+```
+$ conda create \
+    -n landspill -c main -c conda-forge -c barbagroup \
+    python=3.8 geoclaw-landspill
+```
+Once activate the environment, the executable `geoclaw-landspill` should already
+be available.
 
-The easiest way to install *geoclaw-landspill* is through PyPI. 
+### Option 2: use `pip` to install from PyPI or from source
+
+Note, when using the `pip` command, users can always add the `--user` flag to
+install to users' local paths and avoid root privilege. However, if using the
+`--user` flag, users should make sure `pip`'s local `bin` path is in `PATH`.
+
+#### Option 2.1: install from PyPI
+
+To install the package from PyPI. 
 ```
 $ pip install geoclaw-landspill
 ```
 
 We only distribute source tarballs on PyPI due to the requirement of a Fortran
-compiler. Wheels or binary releases of this package are not available.
+compiler. Wheels or binary releases of this package are not available. `pip`
[…]
