### 2021-05-21 784c7d7
setup a default omp thread number to the count of physical cores
addressing the first comment in #20
