+++ conda/meta.yaml
@@ -24,6 +24,8 @@ requirements:
     - scikit-build
   run:
     - python >=3.7
+    - six >=1.15.0
+    - psutil >=5.8.0
     - requests >=2.25.1
     - numpy >=1.19.4
     - matplotlib >=3.3.3
