### 2021-05-21 989559e
add a CMD arg to overwrite the standard output verbosity
addressing the second comment in #20
