+++ gclandspill/__main__.py
@@ -48,6 +48,9 @@ def main():
         "case", action="store", type=pathlib.Path, metavar="CASE",
         help="The path to the target case directory."
     )
+    parser_run.add_argument(
+        '--log-level', dest="log_level", action="store", type=int, default=None,
+        help='Overwrite the log verbosity in the config file `setrun.py` (default: no overwrite)')
     parser_run.set_defaults(func=run)  # set the corresponding callback for the `run` command
 
     # `createnc` command
@@ -311,7 +314,7 @@ def run(args: argparse.Namespace):
     args.output = args.case.joinpath("_output")
 
     # create *.data files, topology files, and hydrological file
-    create_data(args.case, args.output)
+    create_data(args.case, args.log_level, args.output)
 
     # get the Fortran solver binary
     solver = pathlib.Path(gclandspill.__file__).parent.joinpath("bin", "geoclaw-landspill-bin")
