### 2021-05-21 bea628b
bump the conda package version tag
