### 2021-05-21 2addade
update conda package hash and homepage; chage conda search order
