### 2021-05-22 #3114 @piyueh
> Dear all,
>
> I've finished review of the paper and I'd like to congratulate the authors on their effort. Some things could be improved as they might be more valuable to first-timers:
>
>     * [ ]  consider adding a link to Contributing.md inside Readme.md. This might be beneficial as the Github page displays contents of Readme.md by default.
>
>     * [ ]  consider adding a dedicated conda package for your solver. This might simplify the installation of your package.
>
>     * [ ]  consider adding an explicit statement that Windows is not supported (or that Windows used can install it through WSL).
>
>
> A good example that covers all of these is this [JOSS paper](https://joss.theoj.org/papers/10.21105/joss.02696).
>
> My recommendations are _not essential_ and I think paper **can be accepted as is**.

@bzindovic Thanks for your comment. I added the link to CONTRIBUTING.md and the information of supported OS in README.me. Conda package is now available: https://anaconda.org/barbagroup/geoclaw-landspill

The information on how to install the package with `conda` is also available in README.md and installation documentation.

The change has been merged to the master branch and has a tag [v1.0.dev4](https://github.com/barbagroup/geoclaw-landspill/tree/v1.0.dev4).
