### 2021-05-22 barbagroup/geoclaw-landspill-cases#20 @piyueh
Hi @jmansour I have added the features to the code. They are now available in the current master branch with tag [v1.0.dev4](https://github.com/barbagroup/geoclaw-landspill/tree/v1.0.dev4).

1. The default number of OpenMP threads is now the number of physical cores. Added in the commit [784c7d70](https://github.com/barbagroup/geoclaw-landspill/commit/784c7d7026dd278da7479a8f07c60630214ca0a6).
2. The logging verbosity can now be changed with CMD flag `--log-level <level>`. Added in the commit [989559e](https://github.com/barbagroup/geoclaw-landspill/commit/989559e5d8159c87ab8f8de3121a2f23bb905e52). However, it only turns off the log messages of iterations and regridings. It does not disable the warnings, for example, CFL warnings. Those warnings can not be disabled without modifying the Fortran code. So for simulations with rapidly changing flow conditions, the standard output may still be overwhelmed by those warning messages.
