### 2021-05-22 #3114 @piyueh
@dlgeorge Thanks for your comment! Here's our reply:

> optional suggestion: "full shallow water equations?" Maybe omit "full," or if you want to be more descriptive use "nonlinear shallow water equations" or "2D nonlinear shallow water equations"..."over topography?" (last one is probably obvious and too much though). (Unless full is commonly used in this application field?)

I agree ***full*** *shallow-water equations* does sound redundant to people familiar with it. We use the adjective ***full*** because we feel that in the area of pipeline risk management, people sometimes use *shallow-water equations* to mean some other simplified models, like 1D open channel models, diffusive wave approximation, and kinematic wave approximation. These models are still commonly used in related industries. So we use ***full*** to emphasize that our model is not a simplified shallow-water equation.

> "unnecessary AMR code" is really just refinement criteria...in flag2refine, correct?

Yes. It does sound a bit confusing. We removed that item from the readme in the master branch. We just wanted to express that our model does not consider other factors like wind, surge, etc., so we just removed them to save some run time from logical operations.

> maybe deps_install_tests.md and README.md have...somewhat redundant...but not completely the same info? I'm not sure what the standard protocol is for such...but possibly could be improved...maybe it could just all be in README?

Yes, it was somehow redundant at the time of reviewing. However, because now users can install the package using `conda`, the README.md now only has a short description of installing through `conda`. All other information is now in `dep_install_tests.md`.

> agree with @bzindovic: "consider adding a dedicated conda package for your solver," though I personally don't know what details that effort would entail.

We added conda packages: https://anaconda.org/barbagroup/geoclaw-landspill

Thanks!
