### 2021-05-22 barbagroup/geoclaw-landspill-cases#21 @piyueh
Addressed in the original JOSS review issue https://github.com/openjournals/joss-reviews/issues/3114#issuecomment-846341934

Closing it and will reopen it or open a new issue if needed.
