### 2021-05-26 #3114 @kbarnhart
@piyueh thank you for the update, apologies for the delay.

@dlgeorge, @jmansour, @bzindovic as you are able, please revisit this review and indicate whether the authors have addressed all of your concerns. If you think the submission is ready to be accepted, please indicate so in a comment.

@piyueh once all reviewers are satisfied with the submission I will move on to the final steps of the JOSS process.
