### 2021-05-27 #3114 @jmansour
Yep, I've had a look and the authors have made the necessary changes to rectify the minor shortcomings I'd highlighted.

Thanks.
