### 2021-06-07 #3114 @dlgeorge
Hi Katy,

yes, the authors have addressed my (minor) concerns, I think it's ready to
be accepted (sorry, I meant to send this last week).

best,

dave

On Sun, May 30, 2021 at 10:46 AM Katy Barnhart ***@***.***>
wrote:

> Thanks for confirming @jmansour <https://github.com/jmansour>.
>
> @bzindovic <https://github.com/bzindovic> @dlgeorge
> <https://github.com/dlgeorge> please revisit your review and confirm
> whether or not you think this is ready to be accepted.
>
