### 2021-06-08 #3114 @kbarnhart
@piyueh I've made one pull requests with very small changes to the text. On that PR I've also made some notes regarding updates to the references. Please consider this PR and update the references. After you have done so, please test that the article builds correctly here, and presuming it does, do the following:

I can then move forward with recommending that JOSS accept this submission.

@dlgeorge and @bzindovic: Thanks for confirming your recommendation to accept this submission.
