### 2021-06-08 #3114 @piyueh
Thank you @kbarnhart

1. The missing DOIs were added to the references.
2. I couldn't find the dates of the websites of Hydronia and RPS Group, so I just left them with `n.d.`
3. The release tag is `v1.0`: https://github.com/barbagroup/geoclaw-landspill/releases/tag/v1.0
4. The DOI of Zenodo archive: 10.5281/zenodo.4914798 (http://doi.org/10.5281/zenodo.4914798)
